SimpleNeRF

Documentation for SimpleNeRF.

SimpleNeRF.bare_rays

SimpleNeRF.points

SimpleNeRF.render_rays

SimpleNeRF.bare_rays — Method

Get all of the rays in the view in raster scan order

Returns 3x2xB array of (origin, direction) pairs

SimpleNeRF.points — Method

For each ray, compute the points at all timesteps

rays: 3x2xB batch of rays

Returns a batch of points of shape 3xTxB

SimpleNeRF.render_rays — Method

Volumetric rendering given density and color samples along a batch of rays.